Fix light-box
