+++ js/renderGallery.js
@@ -1,20 +1,12 @@
 import renderLightBox from './renderLightBox.js';
 
 function renderGallery(selector, dataType) {
-    const gallery = document.querySelector(selector);
-    const galleryClose = document.querySelector('.light-box .fa-times');
+
     const galleryBack = document.querySelector('.light-box .fa-angle-left');
     const galleryForward = document.querySelector('.light-box .fa-angle-right');
     const imagesArray = document.querySelectorAll(`[data-type="${dataType}"]`);
-    function openOnClick(event) {
-        const images = event.target.parentElement.children;
-        event.target.closest('body').querySelector('.light-box').classList.add('show');
-        return images;
-    }
 
-    function closeOnClick(event) {
-        event.target.closest('body').querySelector('.light-box').classList.remove('show');
-    }
+
     let count = 0;
     function backOnClick(event) {
         count = count - 1;
@@ -44,8 +36,7 @@ function renderGallery(selector, dataType) {
     galleryForward.addEventListener('click', forwardOnClick);
 
 
-    gallery.addEventListener('click', openOnClick);
-    galleryClose.addEventListener('click', closeOnClick);
+
     return;
 }
 export default renderGallery;
